// src/lib/managers/radarr/generated.ts
export interface paths {
// ...
    "/api": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: {
            parameters: {
                query?: never;
                header?: never;
                path?: never;
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description OK */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["ApiInfoResource"];
                    };
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/api/v3/movie": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: {
            parameters: {
                query?: {
                    tmdbId?: number;
                    excludeLocalCovers?: boolean;
                    languageId?: number;
                };
                header?: never;
                path?: never;
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description OK */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "text/plain": components["schemas"]["MovieResource"][];
                        "application/json": components["schemas"]["MovieResource"][];
                        "text/json": components["schemas"]["MovieResource"][];
                    };
                };
            };
        };
        put?: never;
        post: {
            parameters: {
                query?: never;
                header?: never;
                path?: never;
                cookie?: never;
            };
            requestBody?: {
                content: {
                    "application/json": components["schemas"]["MovieResource"];
                };
            };
            responses: {
                /** @description OK */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["MovieResource"];
                    };
                };
            };
        };
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v3/movie/{id}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: {
            parameters: {
                query?: never;
                header?: never;
                path: {
                    id: number;
                };
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description OK */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "text/plain": components["schemas"]["MovieResource"];
                        "application/json": components["schemas"]["MovieResource"];
                        "text/json": components["schemas"]["MovieResource"];
                    };
                };
            };
        };
        put: {
            parameters: {
                query?: {
                    moveFiles?: boolean;
                };
                header?: never;
                path: {
                    id: string;
                };
                cookie?: never;
            };
            requestBody?: {
                content: {
                    "application/json": components["schemas"]["MovieResource"];
                };
            };
            responses: {
                /** @description OK */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["MovieResource"];
                    };
                };
            };
        };
        post?: never;
        delete: {
            parameters: {
                query?: {
                    deleteFiles?: boolean;
                    addImportExclusion?: boolean;
                };
                header?: never;
                path: {
                    id: number;
                };
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description OK */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content?: never;
                };
            };
        };
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
// ...
        MovieResource: {
            /** Format: int32 */
            id?: number;
            title?: string | null;
            originalTitle?: string | null;
            originalLanguage?: components["schemas"]["Language"];
            alternateTitles?: components["schemas"]["AlternativeTitleResource"][] | null;
            /** Format: int32 */
            secondaryYear?: number | null;
            /** Format: int32 */
            secondaryYearSourceId?: number;
            sortTitle?: string | null;
            /** Format: int64 */
            sizeOnDisk?: number | null;
            status?: components["schemas"]["MovieStatusType"];
            overview?: string | null;
            /** Format: date-time */
            inCinemas?: string | null;
            /** Format: date-time */
            physicalRelease?: string | null;
            /** Format: date-time */
            digitalRelease?: string | null;
            /** Format: date-time */
            releaseDate?: string | null;
            physicalReleaseNote?: string | null;
            images?: components["schemas"]["MediaCover"][] | null;
            website?: string | null;
            remotePoster?: string | null;
            /** Format: int32 */
            year?: number;
            youTubeTrailerId?: string | null;
            studio?: string | null;
            path?: string | null;
            /** Format: int32 */
            qualityProfileId?: number;
            hasFile?: boolean | null;
            /** Format: int32 */
            movieFileId?: number;
            monitored?: boolean;
            minimumAvailability?: components["schemas"]["MovieStatusType"];
            isAvailable?: boolean;
            folderName?: string | null;
            /** Format: int32 */
            runtime?: number;
            cleanTitle?: string | null;
            imdbId?: string | null;
            /** Format: int32 */
            tmdbId?: number;
            titleSlug?: string | null;
            rootFolderPath?: string | null;
            folder?: string | null;
            certification?: string | null;
            genres?: string[] | null;
            tags?: number[] | null;
            /** Format: date-time */
            added?: string;
            addOptions?: components["schemas"]["AddMovieOptions"];
            ratings?: components["schemas"]["Ratings"];
            movieFile?: components["schemas"]["MovieFileResource"];
            collection?: components["schemas"]["MovieCollectionResource"];
            /** Format: float */
            popularity?: number;
            /** Format: date-time */
            lastSearchTime?: string | null;
            statistics?: components["schemas"]["MovieStatisticsResource"];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}